+++ river/services/state.py
@@ -1,9 +1,11 @@
 from django.db.models import Q
-from river.models import FORWARD, BACKWARD
+
+from river.models import FORWARD
 from river.models.proceeding import Proceeding
 from river.models.state import State
 from river.utils.error_code import ErrorCode
 from river.utils.exceptions import RiverException
+from .proceeding import ProceedingService
 
 __author__ = 'ahmetdal'
 
@@ -41,39 +43,22 @@ class StateService:
         """
         A state which is not a destination of a transition but can be source of a transition OR not (a destination of a transition and this transition direction is FORWARD)
         """
-        initial_state_candidates = State.objects.filter(
-            Q(transitions_as_source__isnull=False,
-              transitions_as_destination__isnull=True,
-              ) &
-            ~Q(
-                transitions_as_destination__isnull=False,
-                transitions_as_destination__direction=FORWARD,
-            )
-        ).distinct()
-        c = initial_state_candidates.count()
-        if c == 0:
+        states = State.objects.filter(pk__in=ProceedingService.get_initial_proceedings(content_type, field).values_list('meta__transition__source_state', flat=True))
+        if states.count() == 0:
             raise RiverException(ErrorCode.NO_AVAILABLE_INITIAL_STATE, 'There is no available initial state for the content type %s. ' % content_type)
-        elif c > 1:
+        elif states.count() > 1:
             raise RiverException(ErrorCode.MULTIPLE_INITIAL_STATE, 'There are multiple initial state for the content type %s. Have only one initial state' % content_type)
 
-        return initial_state_candidates[0]
+        return states[0]
 
     @staticmethod
     def get_final_states(content_type, field):
+
         """
         A state which is not a source of a transition but can be destination of a transition OR not (a source of a transition and this transition direction is FORWARD)
         """
-        final_states = State.objects.filter(
-            Q(transitions_as_source__isnull=True,
-              transitions_as_destination__isnull=False,
-              ) &
-            ~Q(
-                transitions_as_source__isnull=False,
-                transitions_as_source__direction=FORWARD,
-            )
-        ).distinct()
-        c = final_states.count()
-        if c == 0:
+        proceedings = ProceedingService.get_final_proceedings(content_type, field)
+        if proceedings.count() == 0:
             raise RiverException(ErrorCode.NO_AVAILABLE_FINAL_STATE, 'There is no available final state for the content type %s.' % content_type)
 
-        return final_states
+        return State.objects.filter(pk__in=proceedings.values_list('meta__transition__destination_state', flat=True))